// src/types/settings.ts
// types/settings.ts

export interface SiteSettings {
    site_name: string;
    site_description: string;
    site_keywords: string;
    site_author: string;
    meta_title: string;
    meta_description: string;
    meta_robots: string;
    favicon_url: string;
    logo_url: string;
    google_analytics_id: string;
    facebook_pixel_id: string;
  }